- Add `Execute()` to `TraceRecord`, accordingly, execution functions are added to `TracerHelper` and removed from `ReplayerImpl` to avoid downcasting; - Rename `Replayer::NextTraceRecord()` to `Replayer::Next()`; - Separate implementations of `TraceRecord` and `ReplayerImpl` into different locations; - Multi-threaded replay has a callback to stop at decoding/execution error.

+++ tools/db_bench_tool.cc
@@ -8024,7 +8024,7 @@ class Benchmark {
     }
     s = replayer->Replay(
         ReplayOptions(static_cast<uint32_t>(FLAGS_trace_replay_threads),
-                      static_cast<uint32_t>(FLAGS_trace_replay_fast_forward)));
+                      FLAGS_trace_replay_fast_forward));
     replayer.reset();
     if (s.ok()) {
       fprintf(stdout, "Replay completed from trace_file: %s\n",
